您的提交信息

+++ C++/LeetCode/代码随想录/04_栈与队列/0150.逆波兰表达式求值.cpp
@@ -26,7 +26,7 @@ class Solution
 public:
     int evalRPN(vector<string> &tokens)
     {
-        long result;
+        long long result;
         stack<long> stack;
         for (const string &str : tokens)
         {
